Add explicit uint32_t casts

+++ src/ecs/item_animation_component.cpp
@@ -79,7 +79,7 @@ ItemAnimationComponent::ItemAnimationComponent(SpriteAnimation *animationInfo)
             state.info = Direction::Forward;
             break;
         case AnimationLoopType::Counted:
-            state.info = 0;
+            state.info = static_cast<uint32_t>(0);
             break;
         default:
             VME_LOG_D("Unknown AnimationLoopType");
